[feat] #11 - layout 수정
